Fixed minor edits
Fixed "vs code" to "git should now be installing". Added note that you must install vs code first in order to install git with it configured with vs code as the default editor. I just tried this on a friend's PC and you can't move forward and click "next" on the Git install otherwise.